+++ package.json
@@ -24,6 +24,7 @@
   },
   "homepage": "https://github.com/arsenidze/buckwheat-int20h#readme",
   "dependencies": {
+    "@material-ui/core": "^4.11.2",
     "concurrently": "^5.3.0",
     "cross-env": "^7.0.3",
     "shelljs": "^0.8.4"
